backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-pubsub
python-multipart
firebase-admin
